Updated ej_data_field primary key id to unsigned small int data type

+++ ej_data_value/create.sql
@@ -3,7 +3,7 @@ CREATE TABLE ej_data_value (
 	id INTEGER UNSIGNED NOT NULL PRIMARY KEY AUTO_INCREMENT,
 	year SMALLINT UNSIGNED NOT NULL,
 	value VARCHAR(20) NOT NULL,
-	ej_data_field_id TINYINT UNSIGNED NOT NULL,
+	ej_data_field_id SMALLINT UNSIGNED NOT NULL,
 	block_group_id INTEGER UNSIGNED NOT NULL,
 
 	FOREIGN KEY (ej_data_field_id) REFERENCES ej_data_field (id),
